+++ inst/vosondash/ui/networkGraphsUI.R
@@ -21,10 +21,17 @@ tabItem(tabName = "network_graphs_tab",
                    ),
                    sidebarPanel(width = 12, class = "custom_well_for_controls",
                                 div("Graph Filters", style = "font-weight: bold;", class = "div_inline"),
-                                div(disabled(actionButton("graph_reseed_button", label = icon("refresh"), style = "padding:2px 8px;")), style = "float:right; margin-top:5px; font-size:0.98em;",
+                                div(div(id = "seed", "", class = "div_inline"), disabled(actionButton("graph_reseed_button", label = icon("refresh"), style = "padding:2px 8px;")), style = "float:right; margin-top:5px; font-size:0.98em;",
                                     vpopover(po_reseed_graph()$title, po_reseed_graph()$content)),
-
-                                disabled(checkboxInput("graph_names_check", "Node Labels", FALSE)),
+                                div(""),
+                                div(disabled(checkboxInput("node_index_check", "Node Index", FALSE)), class = "div_inline", style = "margin-right:8px; margin-top:0px;"),
+                                div(disabled(checkboxInput("node_labels_check", "Use Labels", FALSE)), class = "div_inline"),
+                                # conditionalPanel(condition = 'input.node_labels_check',
+                                #                  fluidRow(column(width = 6, shinyjs::disabled(selectInput("node_label_select", label = NULL, 
+                                #                                                         choices = c("None"), selected = NULL, multiple = FALSE))))
+                                # ),
+                                div(""),
+                                
                                 div(disabled(checkboxInput("graph_multi_edge_check", "Multiple Edges", TRUE)), class = "div_inline", style = "margin-right:8px; margin-top:0px;"),
                                 div(disabled(checkboxInput("graph_loops_edge_check", "Loops", TRUE)), class = "div_inline", style = "margin-right:8px; margin-top:0px;"),
                                 div(disabled(checkboxInput("graph_isolates_check", "Isolates", TRUE)), class = "div_inline"),
@@ -33,26 +40,37 @@ tabItem(tabName = "network_graphs_tab",
                                          div(tags$b("Graph Layout"), 
                                              vpopover(po_graph_layout()$title, po_graph_layout()$content), 
                                              style = "margin-bottom:5px;"),
-                                         disabled(selectInput("graph_layout_select", label = NULL, choices = c("Auto", "FR", "KK", "DH",
-                                                                                                                         "LGL", "Graphopt", "DrL", "GEM", "MDS",
-                                                                                                                         "Grid", "Sphere", "Circle", "Star", "Random"),
+                                         disabled(selectInput("graph_layout_select", label = NULL, choices = c("Auto", "FR", "KK", "DH", "LGL", "Graphopt", "DrL", "GEM",
+                                                                                                               "MDS", "Grid", "Sphere", "Circle", "Star", "Random"),
                                                               selectize = TRUE, selected = "Auto"))
                                   ),
                                   column(width = 6,
                                          disabled(sliderInput("graph_spread_slider", "Spread", min = 0.25, max = 2.5, step = 0.1, value = c(1), ticks = FALSE))
                                   )
                                 ),
-                                
+                                conditionalPanel(condition = 'input.graph_layout_select == "FR" | input.graph_layout_select == "Graphopt"',
+                                                 fluidRow(column(width = 6, numericInput(inputId = "graph_niter", "Iterations (niter)", value = 500, min = 1, max = 1000000)))
+                                ),
+                                conditionalPanel(condition = 'input.graph_layout_select == "Graphopt"',
+                                                 fluidRow(column(width = 6, numericInput(inputId = "graph_charge", "Charge", value = 0.001, min = 0.001, max = 1.0, step = 0.001)),
+                                                          column(width = 6, numericInput(inputId = "graph_mass", "Mass", value = 30, min = 1, max = 1000))),
+                                                 fluidRow(column(width = 6, numericInput(inputId = "graph_spr_len", "Spring Length", value = 0, min = 0, max = 1000)),
+                                                          column(width = 6, numericInput(inputId = "graph_spr_const", "Constant", value = 1, min = 1, max = 1000)))
+                                ),
                                 fluidRow(
                                   column(width = 6,
                                          div("Node Size", style = "font-weight: bold;", class = "custom_node_size_div"),
-                                         disabled(selectInput("graph_node_size_degree_select", label = NULL, choices = c("None", "Degree", "Indegree", "Outdegree", "Betweenness", "Closeness"),
+                                         disabled(selectInput("graph_node_size_select", label = NULL, choices = c("None", "Degree", "Indegree", "Outdegree", "Betweenness", "Closeness"),
                                                               multiple = FALSE, selectize = TRUE))
                                   ),
                                   column(width = 6,
                                          disabled(sliderInput("graph_node_size_slider", label = "Multiplier", min = 0.1, max = 15, step = 0.1, value = c(1), ticks = FALSE, animate = FALSE))
                                   )
                                 ),
+                                
+                                checkboxInput('use_vertex_colors_check', 
+                                              div("Node colors from graphml", style = "margin-bottom:5px;")
+                                              , TRUE),
                                 
                                 checkboxInput('expand_categorical_filter_check', 
                                               div(tags$b("Categorical Filter"),